pyproject.toml
[project]
name = "scribe-templates"
version = "0.3.1"
description = "Rapid DOCX generation from Jinja2 templates – YAML + pydantic + Typer."
authors = [
    {name = "Alex"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docxtpl (>=0.20.1,<0.21.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "types-pyyaml (>=6.0.12.20250516,<7.0.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pandas-stubs (>=2.3.0.250703,<3.0.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "hypothesis (>=6.135.32,<7.0.0)",
    "coverage (>=7.9.2,<8.0.0)"
]


[tool.poetry]
packages = [
    { include = "scribe" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
scribe = "scribe.cli.main:app"

[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
hypothesis = "^6.136.6"
coverage = "^7.10.1"
types-pyyaml = "^6.0.12.20250516"
pandas-stubs = "^2.3.0.250703"

[tool.mypy]
python_executable = ".venv/bin/python"
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
show_error_codes = true
warn_unused_ignores = true
warn_return_any = true
disallow_any_unimported = true